requirements.txt
Django==1.11.6
django-bootstrap3==9.0.0
django-crispy-forms==1.9.0
django-filter==2.0.0
djangorestframework==3.9.4
Markdown==3.1.1
Pillow==6.2.2
psycopg2==2.8.4
pytz==2019.3
virtualenv==16.7.9